add roles user, admin, superadmin

+++ api/handlers.py
@@ -7,7 +7,12 @@ from fastapi import HTTPException
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from api.actions.user import _create_new_user, _delete_user, _get_user_by_id, _update_user
+from api.actions.auth import get_current_user_from_token
+from api.actions.user import _create_new_user
+from api.actions.user import _delete_user
+from api.actions.user import _get_user_by_id
+from api.actions.user import _update_user
+from api.actions.user import check_user_permissions
 from api.models import DeleteUserResponse
 from api.models import ShowUser
 from api.models import UpdateUserRequest
@@ -37,25 +42,31 @@ async def create_user(body: UserCreate, db: AsyncSession = Depends(get_db)) -> S
 async def delete_user(
     user_id: UUID,
     db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(get_current_from_token),
+    current_user: User = Depends(get_current_user_from_token),
 ) -> DeleteUserResponse:
-    delete_user_id = await _delete_user(user_id, db)
-    if delete_user_id is None:
+    user_for_deletion = await _get_user_by_id(user_id, db)
+    if user_for_deletion is None:
         raise HTTPException(status_code=404, detail=f"User {user_id} not found")
-    return DeleteUserResponse(delete_user_id=delete_user_id)
+    if not check_user_permissions(
+        target_user=user_for_deletion,
+        current_user=current_user,
+    ):
+        raise HTTPException(status_code=403, detail="Forbidden.")
+    deleted_user_id = await _delete_user(user_id, db)
+    if deleted_user_id is None:
+        raise HTTPException(status_code=404, detail=f"User {user_id} not found")
+    return DeleteUserResponse(deleted_user_id=deleted_user_id)
 
 
 @user_router.get("/", response_model=ShowUser)
 async def get_user_by_id(
-        user_id: UUID,
-        db: AsyncSession = Depends(get_db),
-        current_user: User = Depends(get_current_user_from_token),
+    user_id: UUID,
+    db: AsyncSession = Depends(get_db),
+    current_user: User = Depends(get_current_user_from_token),
 ) -> ShowUser:
     user = await _get_user_by_id(user_id, db)
     if user is None:
-        raise HTTPException(
-            status_code=404, detail=f"User {user_id} not found"
-        )
+        raise HTTPException(status_code=404, detail=f"User {user_id} not found")
     return user
 
 
@@ -69,9 +80,14 @@ async def update_user_by_id(
     updated_user_params = body.dict(exclude_unset=True)
     if updated_user_params == {}:
         raise HTTPException(status_code=422, detail="At least parametr for user update")
-    user = await _get_user_by_id(user_id, db)
-    if user is None:
+    user_for_update = await _get_user_by_id(user_id, db)
+    if user_for_update is None:
         raise HTTPException(status_code=404, detail=f"User {user_id} not found")
+    if user_id != current_user.user_id:
+        if check_user_permissions(
+            target_user=user_for_update, current_user=current_user
+        ):
+            raise HTTPException(status_code=403, detail="Forbidden.")
     try:
         updated_user_id = await _update_user(
             updated_user_params=updated_user_params, session=db, user_id=user_id
